mv_ddr: revise atf part of makefile
This patch is mainly a thorough revision of ATF part of mv_ddr Makefile
without changing its functionality. It adds an option to relocate mv_ddr4 code.
The default location of mv_ddr4 code is in parallel with mv_ddr.
Set MV_DDR4_PATH to a new location of mv_ddr4 code to enable this feature.

Change-Id: Id7e8fc8d038adc4d7d33d24df46e09f9c4fe822d
Signed-off-by: Victor Axelrod <victora@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/33648
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Nadav Haklai <nadavh@marvell.com>