mv_ddr: implement new tx search and centralization and vref algorithm
this patch handles the tx search algorithm which implemented by the
ddr3_tip_ip_training_wrapper function as following:
Training search & read result routine
This function implements the search algorithm
first it calls the function ddr3_tip_ip_training_wrapper_int which triggers the search from l2h and h2l
this function handles rx and tx search cases
in case of rx it only triggers the search (l2h and h2l)
in case of tx there are 3 optional algorithm phases:
phase 1:
it first triggers the search and handles the results as following (phase 1):
each bit, which defined by the search two edges (e1 or VW_L and e2 or VW_H), match on of cases:
 1.	BIT_LOW_UI	0 =< VW =< 31
 2.	BIT_HIGH_UI	32 =< VW =< 63
 3.	BIT_SPLIT_IN	VW_L <= 31 & VW_H >= 32
 4.	BIT_SPLIT_OUT*	VW_H < 32 &  VW_L > 32
note: the VW units is adll taps
phase 2:
only bit case BIT_SPLIT_OUT requires another search (phase 2) from the middle range in two directions h2l and l2h
because only this case is not locked by the search engine in the first search trigger (phase 1).
phase 3:
each subphy is categorized according to its bits definition.
the sub-phy cases are as follows:
 1.BYTE_NOT_DEFINED			the byte has not yet been categorized
 2.BYTE_HOMOGENEOUS_LOW		0 =< VW =< 31
 3.BYTE_HOMOGENEOUS_HIGH		32 =< VW =< 63
 4.BYTE_HOMOGENEOUS_SPLIT_IN		VW_L <= 31 & VW_H >= 32
					or the center of all bits in the byte  =< 31
 5.BYTE_HOMOGENEOUS_SPLIT_OUT	VW_H < 32 &  VW_L > 32
 6.BYTE_SPLIT_OUT_MIX		at least one bits is in split out state and one bit is in other
					or the center of all bits in the byte => 32
after the two phases above a center valid window for each subphy is calculated accordingly:
center valid window = maximum center of all bits in the subphy - minimum center of all bits in the subphy.
now decisions are made in each subphy as following:
all subphys which are homogeneous remains as is
all subphys which are homogeneous low | homogeneous high and the subphy center valid window is less than 32
	mark this subphy as homogeneous split in.
now the bits in the bytes which are BYTE_SPLIT_OUT_MIX needed to be reorganized and handles as following
all bits which are BIT_LOW_UI will be added with 64 adll,
this will hopefully ensures that all the bits in the sub phy can be sampled by the dqs

this patch also handles the centralization algorithm as follwing:
the target of the centralization is to center all dqs in each byte
the centralization is run over several different pattens and finds
the optimal window for all patterns
the centralization algorithm handles two modes: read and write
- in case of read:
the training ip engine driver will execute the search
and will inintiated by the centralization and will result
two cases:
1. the byte has locked and the result will be marked as
   homogeneous low byte
2. the byte has not locked and the result will be marked as
   not defined byte
in the first case the centralization will handle the byte and will
find the center of the byte
in the second case it will return as fail

- in case of write:
all patterns results are saved into data base
and each pattern byte status is read from the training ip driver
all different patterns byte status are compared and in case
of byte which hold results of mixed byte and homogeneous low byte
for different patten the byte with the homogeneous low pattern will
be added with 64 adlls in order to enable its centralization later on.

this patch also handles the vref algorithm for write operation
the vref algorithm searches for the optimal vref
goes from the highst vref to lowest vref and searchs for a lock in
each vref
the vref calls the centralization in each vref point
the total solution for each byte is described as a polygon
which the center of it will result the optimal adll and vref
the vref will read the status of the byte from the centralization
and in case of mixed byte and homogeneous low byte for different vrefs
it will add 64 adlls to the homogeneous low byte and will continue to
find the optimal vref vs adll point for the current byte

Change-Id: I6fb3797feadc9857f94ffd8561c42853eb79d3b3
Signed-off-by: Ofir Fedida <ofedida@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/35011
Reviewed-by: Nadav Haklai <nadavh@marvell.com>
Tested-by: Nadav Haklai <nadavh@marvell.com>