mv_ddr: a38x: fix suspend-to-ram issue
This patch fixes a38x suspend-to-ram issue by preventing the ddr code
to write to the address space higher than 0x3000 used by Linux.
This fix reduces the number of patterns to be loaded for the ddr
training and removes the gap between them.

Important note: this fix supports only single chip select.

Change-Id: I05c20586202c812de861c477dd6095ef4969b9bf
Signed-off-by: Ofir Fedida <ofedida@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40556
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Axelrod <victora@marvell.com>