mv_ddr: a38x: relocate sys_env_dlb_config_ptr_get func
This patch relocates sys_env_dlb_config_ptr_get func to the file, where
it is called and the corresponding dlb config table is defined.
It allows to set the function and the table to be visible in this file only.
The same is done for brd round trip delay array.

Change-Id: I90d57884b7dc6947a0a88107c7a368b7bf8e3547
Signed-off-by: Victor Axelrod <victora@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/41916
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Ofir Fedida <ofedida@marvell.com>