Work around gcc 8.0.1 thinking wr_data may be used uninitialized
gcc 8.0.1 complains about wr_data possibly being used uninitialized
(causing a warning in code built with -Werror, breaking the build).

This is a false positive because wr_data is initialized correctly,
gcc fails to see that anything that would trigger
if(orig_phase >= 1) is covered by the if(orig_phase > 1) and
else if(orig_phase == 1) clauses above.

While this is not very nice, disabling the bogus warning for the
problematic construct allows mv-ddr-marvell to build (and work) with gcc
8.

Signed-off-by: Bernhard Rosenkränzer <bero@lindev.ch>