Allow access to low addresses with gcc 12
gcc 12 added a warning that triggers on access to low addresses. Add a
compile option that allows access to lower addresses.

Add the 'cc_option' macro to avoid the compile option when the compiler 
does not support it.

This fixes build with TF-A. TF-A added a similar fix in commit
dea23e245fb89.

See some more details in
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=105523

Signed-off-by: Baruch Siach <baruch@tkos.co.il>